pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "saint-jerome"
version = "0.1.0"
description = "Discord bot skeleton for structured Bible reference processing."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "discord.py>=2.5.0,<3.0.0",
  "pydantic-settings>=2.0.0",
  "aiosqlite>=0.20.0",
  "tzdata>=2024.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0,<9.0.0",
  "pytest-asyncio>=0.23.0",
  "ruff>=0.3.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
